Add updateTasks module
Update tasks will check if each object already exists,
if it does it updates the existing values with the newly
input values. If a task does not already exists it adds a
randomly generated id to the object and pushes that
object to the tasks array which holds all task objects.