Fix add task pop up function
Previously, the function was called 'apiAddTask'.
Originally, all functionality for the pop up was to be in
this one function. It was decided that splitting the
code into separate function would be cleaner. So, the name
of the function was changed to reflect the small scope
this function will now have.